V3.01 - final changes before swapping phys location
Made a minor change after last push, but am swapping locations, so dont have access to this device, so am pushing again to save these changes.
Added a case during the update graph which, every time 5k points are being plotted, halves the number of points to increase performance for large iteration counts (as otherwise plotting the graph becomes a bottleneck).
Also fixed an issue where going back to main menu (not currently possible due to an issue with menus) would not restart the graph, so restarting the simulation would break the graphing feature